I may have forgotten to save a file for one of the last 2 commits, this is up to date

+++ controllers/quackCon.js
@@ -25,7 +25,7 @@ router.post('/create', async (req,res) => {
 router.delete('/:id/delete', async (req, res) => {
     try{
         const quack = await db.Quack.findById(req.params.id);
-        //if (req.session.currentUser.id === quack.user){
+        if (req.session.currentUser.id === quack.user){
             const delQuack = await db.Quack.findByIdAndDelete(req.params.id);
             delQuack.likes.forEach(async (like) =>{
                 const user = await db.User.findById(like);
@@ -33,7 +33,7 @@ router.delete('/:id/delete', async (req, res) => {
                 user.save();
             })
             res.redirect('/');
-        //}
+        }
 
     } catch (error) {
         console.log(error);
